LibreLibreUp follower mode

+++ xdrip/BluetoothTransmitter/CGM/Libre/Atom/CGMAtomTransmitter.swift
@@ -445,7 +445,7 @@ class CGMAtomTransmitter:BluetoothTransmitter, CGMTransmitter {
         
     }
     
-    func maxSensorAgeInDays() -> Int? {
+    func maxSensorAgeInDays() -> Double? {
         
         return libreSensorType?.maxSensorAgeInDays()
